test: isolate probe fixtures from production entrypoint guards
